app/requirements.txt
streamlit
python
plotly
pandas
numpy
matplotlib.pyplot 
plotly.express
seaborn
numpy
time